libcamera: pipeline: Put single file pipelines into anonymous namespace
Putting a symbol into an anonymous namespace gives its internal linkage.
This is useful for a couple reasons:
  * the compiler can do more optimizations (e.g. inlining);
  * fewer symbols are exported from the DSO;
  * unused symbols can be better diagnosed.

For example, as it turns out, `PipelineHandlerMaliC55::applyScalerCrop()`
is not actually used, which was not diagnosed previously.

Similarly, `ISICameraData::pipe()` is also not used, it is removed.

This could be done with the other pipeline handlers, but is not done,
due to the issues caused by the log category symbols and to avoid
inconsistencies where half of a pipeline handler is in one namespace
while the other half is somewhere else.

Signed-off-by: Barnabás Pőcze <barnabas.pocze@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>