apps: common: Fix flipped error condition
When OptionsParser::childOption encounters an error, it returns nullptr
and the corresponding option.  The check for a nested parent checks for
an error erroneously, leading not only to a wrong behaviour in case
there's no error but also to a segmentation error in case of error due
to `options' being set to nullptr and accessed later.

Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>