apps: cam: Add support for multiple sinks
It would be useful if cam supported output to different sinks when run
with multiple streams.  For example, a processed stream could be
displayed while a raw stream would be sent to files.  Or a processed
stream could be output to PPM files while a raw stream to raw files.

Let's start with replacing the single CameraSession::sink_ with a vector
of sinks_.  We put just the default sink there for now, more
functionality will be added in followup patches.

Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>