apps: cam: Track sink streams
In order to support different sinks for different streams, we must track
the relationship between sinks and streams.  Let's track the
corresponding streams in FrameSink.  Different kinds of sinks use
different approaches to multiple streams; keeping the sinks to handle
that this way is the simplest.

Let's add FrameSink class member and some utility methods for the
purpose.

We still create only one, the default, sink, so all the streams are
added to the default sink unconditionally.  This will be changed in
a followup patch.

Operations on each given sink are applied only for streams the sink
contains.

Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>