apps: cam: Add command line options for stream specific outputs
Now, when the support for multiple outputs is ready, let's add command
line options to actually use it.

To remain backward compatible and to keep the things simple for the
user, the original output related command line options work as before.
For example, simply running

  cam -c1 -C -D

or

  cam ... -s role=viewfinder,... -s role=raw,... -Fimage#

still works.

But the user can also specify a different output for each of the given
streams now, for example:

  cam ... -s role=viewfinder,... --stream-display \
          -s role=raw,... --stream-file=raw#

or

  cam ... -s role=viewfinder,... --stream-file=processed#.ppm \
          -s role=raw,... --stream-file=raw#

To not complicate the implementation more than necessary, it's not
allowed to combine the old and the new output options.

It is also checked that --stream-display is not used more than once (the
KMS output may be used only for a single stream) and that each stream
has at most one output assigned.

Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>