ipa: rkisp1: Add correction for exposure quantization
When WDR regulation is active, it can happen that exposure times get set
to very low values (Sometimes 2-3 lines). This intentionally introduced
underexposure is corrected by the GWDR module. As exposure time is
quantized by lines, the smalles possible change in exposure time now
results in a quite visible change in perceived brightness.

Mitigate that by applying a small global gain to account for the error
introduced by the exposure quantization.

ToDo: This needs perfect frame synchronous control of the sensor to work
properly which is not guaranteed in all cases. It still improves the
behavior with the current regulation.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>