libcamera: pipeline: Add function to clear acquired media devices
Some pipeline handler implementations, such as the simple pipeline
handler, can encounter systems with multiple media devices that can be
supported.

In a subsequent patch, the simple pipeline handler will be updated so
that it will continue to match other media graphs even if the first
media graph it encounters has an invalid media device. In this case,
there would be a mix of invalid and valid media devices in
mediaDevices_.

Add a function to clear mediaDevices_ so that this can be
prevented.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>