pipeline: rkisp1: Add more info when parameter queue fails
When the rkisp1 driver of the running kernel supports less parameter
types than the libcamera build and such a parameter is used, queuing of
the parameter buffer fails with -EINVAL. For the time being we have no
way to query the kernel for supported parameter types at runtime. So at
least print a error message that points the user in that direction.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>