libcamera: formats: Add a helper to check for a raw pixel format
There are several places with the same pattern to check whether a given
pixel format is a raw format:

  return libcamera::PixelFormatInfo::info(pixFmt).colourEncoding ==
         libcamera::PixelFormatInfo::ColourEncodingRAW;

Let's move the corresponding isFormatRaw helper from mali-c55.cpp to
formats.cpp and use it where applicable.  This also avoids a need to
introduce the same helper (or the same pattern) in the followup patches.

Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>