libcamera: software_isp: Add saturation control
Saturation control is added on top of the colour correction matrix.  A
way of saturation adjustment that can be fully integrated into the
colour correction matrix is used.  The control is available only if Ccm
algorithm is enabled.

The control uses 0.0-2.0 value range, with 1.0 being unmodified
saturation, 0.0 full desaturation and 2.0 quite saturated.

The matrix for saturation adjustment is taken from
https://www.graficaobscura.com/matrix/index.html.  The colour correction
matrix is applied before gamma and the given matrix is suitable for such
a case.  Alternatively, the transformation used in libcamera rpi ccm.cpp
could be used.

Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Tested-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>