libcamera: pipeline: rpi: Don't rely on bufferCount
Currently the raspberrypi pipeline handler relies on bufferCount to
decide on the number of buffers to allocate internally and for the
number of V4L2 buffer slots to reserve.

There already exists a procedure for determining the number of buffers
to reserve, so to remove reliance on bufferCount we simply replace the
one instance that it is used in, as well as remove populating it in
generateConfiguration().

Signed-off-by: Nícolas F. R. A. Prado <nfraprado@collabora.com>
Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
Signed-off-by: Sven Püschel <s.pueschel@pengutronix.de>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>