libcamera: pipeline: imx8-isi: Don't rely on bufferCount
Instead of using bufferCount as the number of V4L2 buffer slots to
reserve in the isi pipeline handler, use a reasonably high constant: 16.
Overallocating isn't a problem as buffer slots are cheap. Having too
few, on the other hand, could degrade performance. It is expected that
this number will be more than enough for most, if not all, use cases.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
Signed-off-by: Sven Püschel <s.pueschel@pengutronix.de>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>