libcamera: pipeline: mali-c55: Don't rely on bufferCount
Instead of using bufferCount as the number of V4L2 buffer slots to
reserve in the mali-c55 pipeline handler. Instead of using the previous
bufferCount value of 4 use the higher value of 16 like in other pipeline
handlers to avoid trashing dmabuf mappings if the application cycles
more than 4 buffers.

This makes way for removing bufferCount.

Signed-off-by: Sven Püschel <s.pueschel@pengutronix.de>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>