v4l2: Allocate buffers based on requested count and MinimumRequests
Currently the number of buffers allocated is based on bufferCount, which
is hardcoded to 1. Instead allocate buffers based on the requested count
and on the MinimumRequests property for the camera, which is accessed
through a newly added getter.

This allows the removal of bufferCount.

While at it, fix a typo: s/interval/internal/.

Signed-off-by: Nícolas F. R. A. Prado <nfraprado@collabora.com>
Reviewed-by: Paul Elder <paul.elder@ideasonboard.com>
Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
Signed-off-by: Sven Püschel <s.pueschel@pengutronix.de>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>