Documentation: guides: pipeline-handler: Document internal queue pattern
Pipeline handlers need to implement a common pattern of adding queued
requests to an internal queue and queuing them to the video devices as
the buffer slots there become available.

Add this pattern to the vivid example in the pipeline-handler guide so
it's clear that new pipeline handlers should also implement it.

Signed-off-by: Nícolas F. R. A. Prado <nfraprado@collabora.com>
Signed-off-by: Sven Püschel <s.pueschel@pengutronix.de>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>