config: Look up log color configuration in configuration file
Let's add a configuration file item for logging in color.

The configuration snippet:

  configuration:
    log:
      color: BOOL

In order to use the global configuration access helper for a boolean
value, we must make a template from it.  To make the template visible,
we must implement it in the header file, otherwise undefined reference
error would be reported when linking.

We don't implement envOption helper for boolean here, because we would
have to deal with a specific value interpretation in this particular
case.

Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>