libcamera: pipeline: rkisp1: Detect invalid sensor configurations
If we select a Sensor Format that is larger than the ISP capabilities
the pipeline will not be able to successfully start.

Detect this during validate - and report accordingly, returning
an 'Invalid' state to reflect that we were not able to
reconfigure to adjust to a working state in this instance.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>