libcamera: camera_sensor: Extend reporting in getFormat()
When a pipeline handler asks the CameraSensor for a format, which is not
achievable, it can be very helpful to the user to identify what size was
requested and why it could not be provided within any constrained
maximum size.

This can make otherwise difficult to diagnose user errors visible and
quickly correctable.

Both of the CameraSensorRaw and CameraSensorLegacy are updated
accordingly but that likely means these two classes could soon benefit
from a more common base class to share otherwise identical code.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Reviewed-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>