libcamera: delayed_controls: Change semantics of sequence numbers
In the context of per frame controls the semantics of
DelayedControls::applyControls() are quite difficult to grasp. The
function

void frameStart(int s) {
	delayedCtrls->applyControls(s);
}

seems intuitively wrong as the frame has already started. I think it is
easier to think about what needs to be done for a specific sequence
number. So (assuming a max sensor delay of 2) the actions would be:

- delayedControls.push(n) stores the controls that shall be active on
  frame n
- delayedControls.get(n) returns these controls again
- delayedControls.apply(n) applies the slowest control for frame n and
  does a look back on other controls. So when a frameStart for frame n
occurs, it is time to call delayedControls.apply(n + maxDelay)

Changing these semantics on delayed controls doesn't require much code
change and has the added benefit that we don't run into clamping for
get() on frames < maxDelay.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>

---

Changes in v3:
- Fixed unit tests
- Added error log when get() is called with an illegal sequence
- Added fixup for SimplePipeline

Changes in v2:
- Applied the semantics change to other pipelines as well.