libcamera: pipeline: uvcvideo: Fix `ExposureTime` control handling
The documentation of `ExposureTime` states that its value is ignored if
`ExposureTimeMode` is not manual. This is currently not handled, and if
`ExposureTimeMode` is automatic and `ExposureTime` is set, then the
controls will be rejected, as expected.

To fix that, keep track of the current value of `ExposureTimeMode`, and
only set `V4L2_CID_EXPOSURE_ABSOLUTE` if the current exposure mode is
manual.

However, the current state of `V4L2_CID_EXPOSURE_AUTO` is not retrieved
when streaming starts, thus if the user does not set `ExposureTimeMode`
then it will remain in an "unset" state and all attempts to set
`ExposureTime` will be ignored. This is done for two reasons:

  * the current values of controls cannot be retrieved with libcamera,
    therefore this unset state cannot be observed by the user;
  * the user will have to set `ExposureTimeMode` either way if they
    want to guarantee that setting `ExposureTime` will have any effect.

If `ExposureTimeMode` is not exposed by the camera, but `ExposeTime` is
supported, then the exposure mode is assumed to be manual to allow
changing `V4L2_CID_EXPOSURE_ABSOLUTE`.

To be able to handle requests that set both `ExposureTime{,Mode}`, process
`ExposureTimeMode` first directly in `processControls()`, and store this
new mode in a temporary location. Then have `processControl()` act on this
temporary state, and only persist the temporary state if `setControls()`
on the video device succeeds.

Bug: https://bugs.libcamera.org/show_bug.cgi?id=242
Signed-off-by: Barnabás Pőcze <barnabas.pocze@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>