vimc: Add cheap frame context queue to return metadata
Add to the vimc pipeline handler a cheap frame context queue so that
metadata can be returned. This is useful for getting feedback on testing
setting controls. It is in the pipeline handler as opposed to the IPA as
all controls in vimc are handled by the pipeline handler and not the
IPA.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>