libcamera: ipa_manager: createIPA: Allow matching by IPA name instead of by pipeline
Parts of the software ISP and the ipa_soft_simple.so IPA may be useful for
more then 1 pipeline handler.

Currently createIPA() / IPAManager::module() assume that there is a 1:1
relationship between pipeline handlers and IPAs and IPA matching is done
based on matching the pipe to ipaModuleInfo.pipelineName[].

One way to allow using a single IPA with multiple pipelines would be to
allow the IPA to declare itself compatible with more then one pipeline,
turning ipaModuleInfo.pipelineName[] into e.g. a vector. But the way
ipaModuleInfo is loaded as an ELF symbol requires it to be a simple flat
C-struct.

Instead add an optional ipaName argument to createIPA() which when set
switches things over to matching ipaModuleInfo.name[] allowing pipelines
to request a specific shared IPA module this way.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>