libcamera: controls: Remove common enum prefix
At the moment all enumerators have a common prefix, in many cases
the name of the control, but not always. This is reasonable for
C++ because currently non-scoped enumerations are used, so some
kind of prefix is needed to differentiate common names like `Auto`,
`Manual`, `On`, `Off`, etc.

However, for e.g. language bindings, it might be more desirable to
have access to the the unprefixed name. (This is even the case for
C++ scoped enumerations.)

Currently, both the gstreamer and python bindings have extra code
to strip the common prefix. So instead of doing that separately in
every binding, etc. store the unprefixed name in the source of truth,
the control/property definition yaml files.

This is an API break, but it only affects C++ because gst and py already
strip the common prefix (TODO: recheck again). And in case of C++ it is
only an API break for enum controls where the common prefix is not the
same as the control name:

   * properties::Location
   * properties::draft::ColorFilterArrangement
   * controls::AwbMode
   * controls::AeConstraintMode
   * controls::AeFlickerMode
   * controls::AeMeteringMode
   * controls::AeExposureMode
   * controls::draft::ColorCorrectionAberrationMode
   * TODO: check again

Signed-off-by: Barnabás Pőcze <barnabas.pocze@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>