libcamera: converter_dw100: Use vertex map
Use the new vertexmap class in the dw100 dewarper implementation.

As the dw100 kernel driver does not use the requests API, controls can
not be tied to a buffer. In the case that the dewarper has buffers
queued for which processing has not yet started, applying the dewarp map
would apply to an already queued buffer. Warn when this situation
occurs.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>

---

Changes in v2:
- Added documentation

Changes in v0.9:
- Include requests support

Changes in v0.8:
- Added warning when queue is not empty