libcamera: pipeline: rpi: Queue metadata until completion
The rpi pipeline drops a certain number of initial frames. It uses
`ControlList::clear()` to remove the metadata items of a request that
was populated while processing a frame that is ultimately dropped.
Since the decision to drop a frame only occurs at the very end of the
processing, this stop-gap measure is introduced to delay metadata
completion until the request is actually completed.

Obsoleted by https://patchwork.libcamera.org/cover/23474/.

Signed-off-by: Barnabás Pőcze <barnabas.pocze@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>