meson: Update `gtest` wrap to 1.17.0
Update `gtest` to the latest available version in the wrapdb,
1.17.0, which was released on 2025-04-30. The wrap file has not
been updated since it was added, using version 1.11.0, which
was released more than 4 years ago.

Due to changes in the wrap file, googletest is compiled as a
library instead of just attaching its sources the build targets
using it, so the `default_library=static` option is added to
preserve the same behaviour.

Signed-off-by: Barnabás Pőcze <barnabas.pocze@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>